Move `AnyData` to its own module, in case we want to use it in other places
