+++ internal/core/any_data.rs
@@ -1,0 +1,24 @@
+//! Dynamically-typed data, for runtime-generic code such as [`ClipboardData`](crate::clipboard::Clipboard).
+
+use crate::SharedString;
+
+/// A piece of data of unspecified type. Use the accessor methods to downcast this to a specific type.
+#[derive(Clone)]
+pub struct AnyData {
+    // Eventually this will be something like `Rc<dyn Any>`, but since we only support `SharedString`
+    // for now, this simplifies the implementation.
+    inner: SharedString,
+}
+
+impl AnyData {
+    /// Returns a reference to the inner value if it is of type `T`, or `None` if it isn’t.
+    pub fn as_string(&self) -> Option<SharedString> {
+        Some(self.inner.clone())
+    }
+}
+
+impl From<SharedString> for AnyData {
+    fn from(value: SharedString) -> Self {
+        Self { inner: value }
+    }
+}
