+++ internal/core/lib.rs
@@ -31,6 +31,7 @@ pub use std::thread_local;
 
 pub mod accessibility;
 pub mod animations;
+pub mod any_data;
 pub mod api;
 pub mod callbacks;
 pub mod clipboard;
@@ -66,6 +67,9 @@ pub mod window;
 
 #[doc(inline)]
 pub use string::SharedString;
+
+#[doc(inline)]
+pub use any_data::AnyData;
 
 #[doc(inline)]
 pub use sharedvector::SharedVector;
